+++ addons/l10n_it_edi/__manifest__.py
@@ -7,7 +7,10 @@
     'version': '0.3',
     'depends': [
         'l10n_it',
-        'account_edi'
+        # Although account_edi is a dependency of account_edi_proxy_client,
+        # it is here because it's in the auto-install
+        'account_edi',
+        'account_edi_proxy_client',
     ],
     'auto_install': ['l10n_it', 'account_edi'],
     'author': 'Odoo',
@@ -21,6 +24,8 @@ E-invoice implementation
         'data/account_edi_data.xml',
         'data/invoice_it_template.xml',
         'data/invoice_it_simplified_template.xml',
+        'data/ir_cron.xml',
+        'views/res_config_settings_views.xml',
         'views/l10n_it_view.xml',
         ],
     'demo': [
