+++ addons/l10n_it_edi/data/ir_cron.xml
