+++ addons/l10n_it_edi/models/__init__.py
@@ -3,8 +3,8 @@
 
 from . import res_partner
 from . import res_company
+from . import res_config_settings
 from . import account_invoice
 from . import account_edi_format
-from . import ir_mail_server
 from . import ddt
 from . import mail_template
