+++ addons/l10n_it_edi/models/res_config_settings.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from odoo import api, models, fields, _
 from odoo.exceptions import UserError
