+++ addons/l10n_it_edi/tests/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from . import test_ir_mail_server
-from . import test_res_partner
+from . import common
+from . import test_edi_export
+from . import test_edi_import
+from . import test_edi_reverse_charge
