+++ addons/l10n_it_edi/tests/expected_xmls/IT00470550013_basis.xml
