+++ addons/l10n_it_edi/tests/expected_xmls/reverse_charge_bill.xml
