+++ addons/l10n_it_edi/views/res_config_settings_views.xml
