--- addons/l10n_it_edi_sdicoop/models/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from . import account_invoice
-from . import account_edi_format
-from . import res_config_settings
