--- addons/l10n_it_edi_sdicoop/tests/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from . import test_edi_xml
-from . import test_edi_reverse_charge_xml
